Divide my code into modules
