+++ index.html
@@ -63,9 +63,6 @@
         </footer>
     </div>
 
-
-
-
-    <script src="./index.js"></script>
+    <script type="module" src="./index.js"></script>
 </body>
 </html>
